Moved NewsItems into its own app called Comms.
