+++ src/core/forms.py
@@ -52,19 +52,6 @@ class EditKey(forms.Form):
         return cleaned_data
 
 
-class NewsItemForm(forms.ModelForm):
-
-    image_file = forms.FileField(required=False)
-
-    def __init__(self, *args, **kwargs):
-        super(NewsItemForm, self).__init__(*args, **kwargs)
-        self.fields['body'].widget = SummernoteWidget()
-
-    class Meta:
-        model = models.NewsItem
-        exclude = ('content_type', 'object_id', 'posted', 'posted_by', 'large_image_file')
-
-
 class JournalContactForm(forms.ModelForm):
 
     class Meta:
