+++ src/press/models.py
@@ -52,7 +52,7 @@ class Press(models.Model):
 
     random_featured_journals = models.BooleanField(default=False)
     featured_journals = models.ManyToManyField('journal.Journal', blank=True, null=True)
-    carousel_news_items = models.ManyToManyField('core.NewsItem', blank=True, null=True)
+    carousel_news_items = models.ManyToManyField('comms.NewsItem', blank=True, null=True)
 
     password_reset_text = models.TextField(blank=True, null=True)
     registration_text = models.TextField(blank=True, null=True)
